chore: extract out search
