+++ src/components/pages/ImportsLayout.tsx
@@ -1,23 +1,7 @@
-import { useQuery } from "@tanstack/react-query";
-import { ChevronsUpDown } from "lucide-react";
 import React, { useState } from "react";
 
-import {
-  Command,
-  CommandEmpty,
-  CommandGroup,
-  CommandInput,
-  CommandItem,
-  CommandList,
-} from "@/components/ui/command";
+import FacilitySelector from "@/components/shared/FacilitySelector";
 import { NavTabs } from "@/components/ui/nav-tabs";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
-import userApi from "@/types/user/userApi";
-import query from "@/Utils/request/query";
 
 export type ImportTabId =
   | "users"
@@ -35,11 +19,6 @@ export type ImportTabId =
 interface ImportsLayoutProps {
   activeTab: ImportTabId;
   children: React.ReactNode;
-}
-
-interface FacilityOption {
-  id: string;
-  name: string;
 }
 
 const getTabConfig = () => [
@@ -105,17 +84,6 @@ export default function ImportsLayout({
   children,
 }: ImportsLayoutProps) {
   const [selectedFacilityId, setSelectedFacilityId] = useState<string>("");
-  const [open, setOpen] = useState(false);
-
-  const {
-    data: facilities = [],
-    isLoading: loadingFacilities,
-    error: facilityError,
-  } = useQuery({
-    queryKey: ["currentUser", "facilities"],
-    queryFn: query(userApi.currentUser),
-    select: (data: { facilities?: FacilityOption[] }) => data.facilities ?? [],
-  });
 
   const tabs = getTabConfig();
   const requiresFacility =
@@ -134,64 +102,10 @@ export default function ImportsLayout({
     <div className="max-w-5xl mx-auto px-4 py-8">
       <div>
         <div className="px-6 pt-4 space-y-4">
-          <div className="flex flex-col gap-2">
-            <label className="text-sm font-medium text-gray-700">
-              Select Facility
-            </label>
-            <Popover open={open} onOpenChange={setOpen}>
-              <PopoverTrigger asChild>
-                <button
-                  type="button"
-                  disabled={loadingFacilities}
-                  className="flex h-9 w-full items-center justify-between rounded-md border border-gray-200 bg-white px-3 py-2 text-sm shadow-xs disabled:cursor-not-allowed disabled:opacity-50"
-                >
-                  <span
-                    className={
-                      selectedFacilityId ? "text-gray-950" : "text-gray-500"
-                    }
-                  >
-                    {loadingFacilities
-                      ? "Loading facilities..."
-                      : selectedFacilityId
-                        ? facilities.find((f) => f.id === selectedFacilityId)
-                            ?.name
-                        : "Select a facility"}
-                  </span>
-                  <ChevronsUpDown className="size-4 shrink-0 opacity-50" />
-                </button>
-              </PopoverTrigger>
-              <PopoverContent className="p-0" align="start">
-                <Command>
-                  <CommandInput
-                    placeholder="Search facilities..."
-                    className="outline-hidden border-none ring-0 shadow-none"
-                    autoFocus
-                  />
-                  <CommandList>
-                    <CommandEmpty>No facilities found.</CommandEmpty>
-                    <CommandGroup>
-                      {facilities.map((facility) => (
-                        <CommandItem
-                          key={facility.id}
-                          value={facility.name}
-                          onSelect={() => {
-                            setSelectedFacilityId(facility.id);
-                            setOpen(false);
-                          }}
-                          className="cursor-pointer"
-                        >
-                          {facility.name}
-                        </CommandItem>
-                      ))}
-                    </CommandGroup>
-                  </CommandList>
-                </Command>
-              </PopoverContent>
-            </Popover>
-            {facilityError && (
-              <p className="text-sm text-red-600">Unable to load facilities</p>
-            )}
-          </div>
+          <FacilitySelector
+            value={selectedFacilityId}
+            onSelect={setSelectedFacilityId}
+          />
           <NavTabs
             tabs={tabs.map((tab) => ({
               key: tab.id,
